Add `trimmed_git_prompt` feature to shell prompt
In attempt to use screen space better we want to reduce length of the
current git branch name we show in the shell prompt. We want to trim if
its length greater than a predefined amount.

The implementation is not the best, as we need to work with the output
of `__fish_git_prompt`, which in its turn configured to output a colored
string.

So, we count a "pure" string length, compare and then cut beginning of
the prompt and its end before we print it to the user.