Use PREFIX+\ and PREFIX+- both in tmux and vim
We want to use the same buttons to split windows vertically and
horizontally in tmux and vim. The only difference is the prefix:

- for tmux it's CTRL+A
- for vim it's CTRL+W

Note: we had to switch to `\` character from `|` because it was
impossible to configure vertical bar shortcut in vim for some reason.
See more here: https://stackoverflow.com/questions/52175573/how-do-you-remap-c-ctrl-bar-in-vim