Add vimrc.ftdetect with detecting filetypes
When we added `.dev` (and `.build`) extensions to `Dockerfile` in one of
ther projects, we noticed that Vim lost syntax highlighting, so we want
to autodetect dockerfiles by `Dockerfile` prefix in the filename.

In future, we might add more filetypes to this list, so it's better to
have a separate configuration file.