Install 'fzf-wordnet.vim' plugin
With this plugin we can use Wordnet application (install it separately!)
to find word completions or look up the definitions in the Wordnet's
vocabulary.

We configured to use <CTRL-s> in insert mode to search for the word
under the cursor.

As an extra work - we also renamed `vimrc.bundles` to more correct
`vimrc.plugins`, as this is the list of plugins to be installed.