Remove some redundancy in .vimrc
We used to use `if filereadable(...)` construction in the .vimrc file to
avoid loading non-existing files. That's silly, as all of these files
are in the same repo, so they must exist (or not – but then remove them
from .vimrc) in the same time period as .vimrc existence.

The `if filereadable` construction makes sense only for checking if
local version of .vimrc exists and then load it (as it is outside of git
repository).