Add elgato.sh – lights automation script (initial)
The script supposed to run in background and watch for special macOS
events that appear in the logs when webcam goes on or off. It
automatically switches on or off Elgato Lights (configured in the
DEVICES variable).

It's an initial version, quite raw and will be rewritten to something
more appropritate later. Though, it already does the minimum: controls
the available lights when camera starts or stops shooting!