Update .gitconfig and .gitconfig.local
It's better to bring signingkey and other commit signature related
options to the main configuration file, as it's easy to forget to unwrap
.gitconfig.local.example on a new system.

As to leave some ideas for the .gitconfig.local, we moved rarely used
`recent` alias definition there. It won't make too harm if we forget to
use it (however an example and data preserved).