Rewrite fish's paths.fish configuration and...
...delete extra files (asdf.fish and homebrew.fish), because their
settings moved to the new paths.fish one.

The main issue we were trying to solve with these changes is the
messiness in the PATH variable after we enter tmux session. Due to the
way how Homebrew (and ASDF) requires us to set them up (by sourcing or
eval-ing some commands they provide), we can hardly predict the order of
the actual paths in the PATH.

This happens also due to the fact that macOS runs
/usr/libexec/path_helper during login (tmux executes shell in the login
mode by default). It reads /etc/paths and /etc/paths.d, then combines it
with the current PATH and sets it back. fish, in its turn, already set
some paths when it loaded with the terminal, in tmux it runs similar
sequence of commands again, and including what path_helper did and
evaluation of `brew shellenv`, we get a mess in the order of PATH.

Basically, with the previous configuration we get this:

1. When terminal is just opened (not in a tmux session), our top
   priority PATH directory set to ASDF's shims one, and then Homebrew
   binaries directories went.
2. When we ran `tbase` (our tmux wrapper) command and enter tmux
   session, Homebrew's directories got to the top.
3. This behavior shadowed some of ASDF-configured apps. For example,
   `node` will be used from Homebrew despite what version is set in the
   .tool-versions file.

With the new approach, we set PATH only if fish is loaded in login mode.

P.S. Also, there are a couple of minor clean ups and indentation fixes.