Start using mason.nvim and mason-lspconfig.nvim
These two plugins should simplify management of the Neovim Language
Servers for a great variety of languages: it supports semi-automatic
installation, updates, and other nice things.

`nvim-lspconfig.lua` config is mostly replaced by the new
`mason-lspconfig.lua` – it mostly repeats what has been configured in
the old version, but cleaned, updated settings for diagnostics floating
window, and disabled (commented) some of the nvim-lspconfig features we
didn't use that much (but there is a chance to use them in future or in
other languages).