Add Mason, Telescope, Mini.nvim, and other...
...plugins and keybindings.

As for the keys, we try to use which-key plugin which provides a nice
way of mapping (and later info window while using Neovim with inline
help).

Mini.nvim might replace some of the previously used plugins, such as:

- tpope/vim-commentary
- tpope/vim-fugitive
- tpope/vim-obsession
- csexton/trailtrash.vim
- Wansmer/langmapper.nvim

We have to try to adapt to mini's settings and abilities.