Re-start Neovim's (and partially, Vim's) configs
During the recent experiements with new Neovim plugins and
configurations, we thought that it would be interesting (and, probably,
useful) to replace Vim's Plug with modern plugin manager (like Lazy), as
well as to cleanup some old Vim's settings that we do not use.

So... This is the beginning of the job: we have removed most of the old
Lua-files used by Neovim, returned some key bindings (and will use them
both in Vim and Neovim), made some cleanups to the .vimrc.

Let's how it will end up.