Add Gatoglow: successor of `elgato-lights.sh`
It was a long-standing idea to rewrite (sometimes) cumbersome Bash
script with Elixir and use GenServer to monitor, log, and react to log
events.

Seems, it works!

Though, it was not that straightforward as you might think...

Zombie processes
---

We had to solve an issue of zombie processes (we do not want to leave
`log stream ...` command running behind if BEAM VM where our controller
script is running crashed).

So, we learned about Port module, and introduced `scruffy.sh` wrapper
script that handles situations like this, and we run the `log stream`
through this wrapper.

Supervisors and long-running system
---

We were also not sure how it is going to work with Supervisors and
GenServers in the script; seems Process.sleep(:infinity) does the job

macOS, launchctl, and ASDF
---

Another challenge was to find a way how to execute our Elixir-script via
launchctl system, because it runs out of user session and doesn't have
default environment that is available in the shell. Plus, Elixir is not
system-wide installed - we use ASDF for that. The solution is in the
proposed gatoglow.plist file. Just don't forget to load it:

```shell
$ copy gatoglow.plist ~/Library/LaunchAgents/
$ launchctl load ~/Library/LaunchAgents/gatoglow.plist
```

P.S. Last but not least: macOS will prompt for permission to let
beam.smp process do network requests. Allow it. But for the first time
it may leave an error record in the log file.