Use (almost) golden ratio for tmux pane sizing
The old 66% we chose to be used by the main pane in tmux and being bound
to the m/M commands, can be (probably) improved.

We decided to try to use https://en.wikipedia.org/wiki/Golden_ratio. It
says that that the main pane should be ~61.8%, so we went with 62 (as we
cannot use floating point numbers).

Let's see how golden ratio tmux session would look like. Initial testing
is promising!