Add tmux popups (!) for IEx and Fish sessions
As it says here - https://www.youtube.com/watch?v=7BP9iWiKx8Q - popups
seem to be really underrated feature of a tmux. It's easy and nice to
open an extra window while currently working with something else, just
in the context of that, without interrupting the work and moving to
another pane or window.

Here, we have added a couple of helpers that we hope to use in the
future.