Replace nvim-cmp with blink.cmp
LazyVim uses blink.cmp as its autocompletion plugins, so why not to try
it? Especially, that it goes only with one package, instead of a whole
family of nvim-cmp subplugins.

There is a little chance that we're losing flexibility of configurations
for autocompletion, but... One less config file!

Just 1 plugin instead of 5:

- saghen/blink.cmp

vs.

- hrsh7th/nvim-cmp
- hrsh7th/cmp-nvim-lsp
- hrsh7th/cmp-buffer
- hrsh7th/cmp-path
- hrsh7th/cmp-nvim-lsp-signature-help